refactor: update temperature handling to use Double instead of Int for better precision

+++ WeatherApp/Core/Formatting/TemperatureFormatter.swift
@@ -7,10 +7,10 @@
 
 
 enum TemperatureFormatter {
-    static func format(celsius: Int, to unit: TemperatureDisplayUnit) -> String {
-        let value: Int = switch unit {
+    static func format(celsius: Double, to unit: TemperatureDisplayUnit) -> String {
+        let value: Double = switch unit {
         case .celsius: celsius
-        case .fahrenheit: Int((Double(celsius) * 9.0/5.0 + 32.0).rounded())
+        case .fahrenheit: (Double(celsius) * 9.0/5.0 + 32.0).rounded()
         }
         return "\(value)°\(unit.suffix)"
     }
